moved santizeName to lib/validation.ts and imported to navbar.tsx and schoolvolunteer.tsx
