+++ src/lib/validation.ts
@@ -4,3 +4,16 @@ const STRICT_EMAIL_REGEX =
 export function isStrictEmail(emailInput: string): boolean {
   return STRICT_EMAIL_REGEX.test(emailInput);
 }
+
+export function sanitizeName(nameInput: string): string {
+  return nameInput
+    .split("")
+    .map((character) => {
+      const characterCode = character.charCodeAt(0);
+      return characterCode <= 31 || characterCode === 127 ? " " : character;
+    })
+    .join("")
+    .replace(/\s+/g, " ")
+    .trim()
+    .slice(0, 100);
+}
